Extract Swatch into a separate file
